one duration vocabulary, shared rather than owned by a section
